requirements.txt
requests
langchain
langchain_community
openai
python-dotenv
